requirements.txt
pandas
requests
boto3
psycopg2-binary
PyYAML
SQLAlchemy
phonenumbers
tabula-py